refactor(zk-explorer): Begin consolidating data fetching functions
