--- apps/zk-explorer/server/index.ts
@@ -1,16 +1,0 @@
-export { default as getProjectCircuitsData } from './getProjectCircuitsData';
-export type {
-  FileType,
-  FileTreeItem,
-  GetProjectCircuitDataReturnType,
-} from './getProjectCircuitsData';
-export { default as getProjectGitHubInfoContainerData } from './getProjectGitHubInfoContainerData';
-export type { ProjectGitHubInfoContainerDataType } from './getProjectGitHubInfoContainerData';
-export { default as getProjectHeaderContainerData } from './getProjectHeaderContainerData';
-export { default as getProjectSummaryData } from './getProjectSummaryData';
-export type {
-  ContributionListItem,
-  ProjectTrustedSetupItem,
-} from './getProjectTrustedSetupData';
-export { default as getProjectTrustedSetupData } from './getProjectTrustedSetupData';
-export { default as getRelatedProjectsContainerData } from './getRelatedProjectsContainerData';
