--- apps/zk-explorer/server/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as getGitHubLanguageColors } from './getGitHubLanguageColors';
