+++ apps/zk-explorer/tsconfig.json
@@ -16,7 +16,16 @@
         "name": "next"
       }
     ],
-    "types": ["jest", "node"]
+    "types": [
+      "jest",
+      "node"
+    ],
+    "baseUrl": ".",
+    "paths": {
+      "@components/*": [
+        "components/*"
+      ],
+    }
   },
   "include": [
     "**/*.ts",
